Move getting started to entirely new document
The current instructions have several pain points which can lead to a
failed deployment which in turn will be hard to debug.

The new guide should be simpler, easier to follow and should actually
work.

The details in the README are rewritten to add context and flavour for
power users, while most people are expected to use the getting started
guide when experimenting.

We also add istio install config specifically for the getting started
guide since the `hack/istio-config-x.yaml` files are also used for e2e
tests; we don't want a change to the e2e test to break the docs, or vice
versa.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>