use specialised function for decoding trust bundles
Importantly, this targets cert-manager's default (master) branch. The
`beta.0` tag which appears in go.mod comes from the `1.16.0-beta.0` tag
being the last one present on master - `v1.16.1` and `v1.16.2` appear on
the release-1.16 branch and so aren't present on master.

A later release of istio-csr (presumably after we tag a beta for
cert-manager 1.17) can target a tagged version.

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>